adding condition

+++ service-scripts/App-MSA.pl
@@ -489,18 +489,20 @@ sub process_fasta
     # Reroot the fasttree nwk file using midpoint rooting, then delete the fasttree nwk
     #
     my $fasttree_nwk = "$work_dir/${prefix}_fasttree.nwk";
-    my $midpoint_tmp = "$work_dir/${prefix}.nwk";
-    my $midpoint_nwk = "$work_dir/${prefix}_midpoint.nwk";
-    my @midpoint_cmd = ("p3x-reformat-tree","--midpoint","-f","newick","-o",$midpoint_tmp,"-i",$fasttree_nwk);
-    print "midpoint root command: @midpoint_cmd\n";
-    run_cmd(\@midpoint_cmd);
-    unlink $fasttree_nwk;
-    # rename midpoint file to same name as fasttree file
-    # ensures MSA tree viewer does not break
-    # terminate if rename doesn't work?:or die "Unable to rename: $!"
-    rename($midpoint_nwk,$fasttree_nwk);
-
-    #
+    # make sure initial nwk file exists
+    if (-e $fasttree_nwk) {
+        my $midpoint_tmp = "$work_dir/${prefix}.nwk";
+        my $midpoint_nwk = "$work_dir/${prefix}_midpoint.nwk";
+        my @midpoint_cmd = ("p3x-reformat-tree","--midpoint","-f","newick","-o",$midpoint_tmp,"-i",$fasttree_nwk);
+        print "midpoint root command: @midpoint_cmd\n";
+        run_cmd(\@midpoint_cmd);
+        unlink $fasttree_nwk;
+        # rename midpoint file to same name as fasttree file
+        # ensures MSA tree viewer does not break
+        # terminate if rename doesn't work?:or die "Unable to rename: $!"
+        rename($midpoint_nwk,$fasttree_nwk);
+    }
+        #
     # Copy output to the workspace.
     #
     my @output_suffixes = (
